Deprecate plot output, move functions to fitting scripts
